Remove no longer needed callout and refactor stuff a bit
