+++ x-pack/plugins/upgrade_assistant/public/application/components/overview/observe_step/_index.scss
@@ -1,1 +1,1 @@
-@import 'deprecation_logging_toggle';
+@import 'deprecation_logging_toggle/deprecation_logging_toggle';
