+++ x-pack/plugins/upgrade_assistant/public/application/components/types.ts
@@ -46,14 +46,3 @@ export enum TelemetryState {
 }
 
 export type EsTabs = 'cluster' | 'indices';
-
-export interface DeprecationLoggingPreviewProps {
-  isEnabled: boolean;
-  isLoading: boolean;
-  isUpdating: boolean;
-  fetchError: ResponseError | null;
-  updateError: ResponseError | undefined;
-  resendRequest: () => void;
-  toggleLogging: () => void;
-  hasLoggerDeprecationWarning: boolean;
-}
